refactor: parse link safes as a util function
